pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantflow-financial-suite"
version = "1.0.0"
description = "AI-Augmented Valuation & Dashboarding for Public Companies"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "yfinance>=0.2.12",
    "beautifulsoup4>=4.11.0",
    "requests>=2.28.0",
    "openpyxl>=3.1.0",
    "matplotlib>=3.6.0",
    "streamlit>=1.28.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "jupyter>=1.0.0",
]
viz = [
    "plotly>=5.13.0",
    "seaborn>=0.12.0",
    "dash>=2.14.0",
]
ml = [
    "scikit-learn>=1.2.0",
    "scipy>=1.10.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/quantflow-financial-suite"
Repository = "https://github.com/yourusername/quantflow-financial-suite"
Documentation = "https://quantflow-docs.readthedocs.io"

[tool.setuptools.packages.find]
where = ["."]
include = ["quantflow*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"